+++ main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 
-	"github.com/poseidon/terraform-provider-matchbox/matchbox"
+	"github.com/poseidon/terraform-provider-matchbox/internal/matchbox"
 )
 
 func main() {
